@@ -147,26 +152,25 @@ endmacro()
 if(APPLE)
     target_compile_options(
         astyle PRIVATE -W -Wall -fno-rtti -fno-exceptions -std=c++11 -stdlib=libc++)
-    set_linker_options("")
+    set_linker_options()
 elseif(NOT WIN32)   # Linux
     target_compile_options(astyle PRIVATE -Wall -fno-rtti -fno-exceptions -std=c++11)
-    set_linker_options("-s")
-elseif(BORLAND)     # Release must be explicitely requested for Borland
-    target_compile_options(astyle PRIVATE -w -x-)   # Cannot use no-rtti (-RT-)
-    set_linker_options("")
+    set_linker_options()
 elseif(MINGW)
-    target_compile_options(astyle PRIVATE -Wall -Wextra -fno-rtti -fno-exceptions -std=c++0x)
-    set_linker_options("-s")
+    target_compile_options(astyle PRIVATE -Wall -Wextra -fno-rtti -fno-exceptions -std=c++11)
+    set_linker_options()
+elseif(BORLAND)     # Release must be explicitely requested for Borland
+    target_compile_options(astyle PRIVATE -q -w -x-)   # Cannot use no-rtti (-RT-)
+    set_linker_options()
 endif()
 
 # Display build information
 if(BUILD_JAVA_LIBS)
-    message("CMAKE_BUILD_TYPE is Java ${CMAKE_BUILD_TYPE} ${SOLIBVER}")
+    message("## CMAKE_BUILD_TYPE is Java ${CMAKE_BUILD_TYPE}")
 elseif(BUILD_SHARED_LIBS)
-    message("CMAKE_BUILD_TYPE is Shared ${CMAKE_BUILD_TYPE} ${SOLIBVER}")
+    message("## CMAKE_BUILD_TYPE is Shared ${CMAKE_BUILD_TYPE}")
 elseif(BUILD_STATIC_LIBS)
-    message("CMAKE_BUILD_TYPE is Static ${CMAKE_BUILD_TYPE}")
+    message("## CMAKE_BUILD_TYPE is Static ${CMAKE_BUILD_TYPE}")
 else()
-    message("CMAKE_BUILD_TYPE is Executable ${CMAKE_BUILD_TYPE}")
+    message("## CMAKE_BUILD_TYPE is Executable ${CMAKE_BUILD_TYPE}")
 endif()
-set(CMAKE_VERBOSE_MAKEFILE ON)
