@@ -187,12 +187,7 @@ bool sortOnName(const string* a, const string* b)
 void ASResource::buildAssignmentOperators(vector<const string*>* assignmentOperators)
 {
 	const size_t elements = 15;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		assignmentOperators->reserve(elements);
-		reserved = true;
-	}
+	assignmentOperators->reserve(elements);
 
 	assignmentOperators->emplace_back(&AS_ASSIGN);
 	assignmentOperators->emplace_back(&AS_PLUS_ASSIGN);
@@ -225,12 +220,7 @@ void ASResource::buildAssignmentOperators(vector<const string*>* assignmentOpera
 void ASResource::buildCastOperators(vector<const string*>* castOperators)
 {
 	const size_t elements = 5;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		castOperators->reserve(elements);
-		reserved = true;
-	}
+	castOperators->reserve(elements);
 
 	castOperators->emplace_back(&AS_CONST_CAST);
 	castOperators->emplace_back(&AS_DYNAMIC_CAST);
@@ -250,12 +240,7 @@ void ASResource::buildCastOperators(vector<const string*>* castOperators)
 void ASResource::buildHeaders(vector<const string*>* headers, int fileType, bool beautifier)
 {
 	const size_t elements = 25;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		headers->reserve(elements);
-		reserved = true;
-	}
+	headers->reserve(elements);
 
 	headers->emplace_back(&AS_IF);
 	headers->emplace_back(&AS_ELSE);
@@ -335,15 +320,10 @@ void ASResource::buildIndentableHeaders(vector<const string*>* indentableHeaders
 void ASResource::buildIndentableMacros(vector<const pair<const string, const string>* >* indentableMacros)
 {
 	const size_t elements = 10;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		indentableMacros->reserve(elements);
-		reserved = true;
-	}
+	indentableMacros->reserve(elements);
 
 	// the pairs must be retained in memory because of pair pointers
-	typedef pair<const string, const string> macro_pair;
+	using macro_pair = pair<const string, const string>;
 	static const macro_pair macros[] =
 	{
 		// wxWidgets
@@ -372,12 +352,7 @@ void ASResource::buildIndentableMacros(vector<const pair<const string, const str
 void ASResource::buildNonAssignmentOperators(vector<const string*>* nonAssignmentOperators)
 {
 	const size_t elements = 15;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		nonAssignmentOperators->reserve(elements);
-		reserved = true;
-	}
+	nonAssignmentOperators->reserve(elements);
 
 	nonAssignmentOperators->emplace_back(&AS_EQUAL);
 	nonAssignmentOperators->emplace_back(&AS_PLUS_PLUS);
@@ -408,12 +383,7 @@ void ASResource::buildNonAssignmentOperators(vector<const string*>* nonAssignmen
 void ASResource::buildNonParenHeaders(vector<const string*>* nonParenHeaders, int fileType, bool beautifier)
 {
 	const size_t elements = 20;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		nonParenHeaders->reserve(elements);
-		reserved = true;
-	}
+	nonParenHeaders->reserve(elements);
 
 	nonParenHeaders->emplace_back(&AS_ELSE);
 	nonParenHeaders->emplace_back(&AS_DO);
@@ -468,13 +438,7 @@ void ASResource::buildNonParenHeaders(vector<const string*>* nonParenHeaders, in
 void ASResource::buildOperators(vector<const string*>* operators, int fileType)
 {
 	const size_t elements = 50;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		operators->reserve(elements);
-		reserved = true;
-	}
-
+	operators->reserve(elements);
 
 	operators->emplace_back(&AS_PLUS_ASSIGN);
 	operators->emplace_back(&AS_MINUS_ASSIGN);
@@ -539,12 +503,7 @@ void ASResource::buildOperators(vector<const string*>* operators, int fileType)
 void ASResource::buildPreBlockStatements(vector<const string*>* preBlockStatements, int fileType)
 {
 	const size_t elements = 10;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		preBlockStatements->reserve(elements);
-		reserved = true;
-	}
+	preBlockStatements->reserve(elements);
 
 	preBlockStatements->emplace_back(&AS_CLASS);
 	if (fileType == C_TYPE)
@@ -584,12 +543,7 @@ void ASResource::buildPreBlockStatements(vector<const string*>* preBlockStatemen
 void ASResource::buildPreCommandHeaders(vector<const string*>* preCommandHeaders, int fileType)
 {
 	const size_t elements = 10;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		preCommandHeaders->reserve(elements);
-		reserved = true;
-	}
+	preCommandHeaders->reserve(elements);
 
 	if (fileType == C_TYPE)
 	{
@@ -628,12 +582,7 @@ void ASResource::buildPreCommandHeaders(vector<const string*>* preCommandHeaders
 void ASResource::buildPreDefinitionHeaders(vector<const string*>* preDefinitionHeaders, int fileType)
 {
 	const size_t elements = 10;
-	static bool reserved = false;
-	if (!reserved)
-	{
-		preDefinitionHeaders->reserve(elements);
-		reserved = true;
-	}
+	preDefinitionHeaders->reserve(elements);
 
 	preDefinitionHeaders->emplace_back(&AS_CLASS);
 	if (fileType == C_TYPE)
@@ -693,10 +642,10 @@ const string* ASBase::findHeader(const string& line, int i,
 		// the following accessor definitions are NOT headers
 		// goto default; is NOT a header
 		// default(int) keyword in C# is NOT a header
-		else if ((header == &AS_GET
-		          || header == &AS_SET
-		          || header == &AS_DEFAULT)
-		         && (peekChar == ';' || peekChar == '(' || peekChar == '='))
+		if ((header == &AS_GET
+		        || header == &AS_SET
+		        || header == &AS_DEFAULT)
+		        && (peekChar == ';' || peekChar == '(' || peekChar == '='))
 			break;
 		return header;
 	}
