@@ -185,12 +185,12 @@ enum LineEndFormat
 class ASSourceIterator
 {
 public:
-	ASSourceIterator() {}
-	virtual ~ASSourceIterator() {}
+	ASSourceIterator() = default;
+	virtual ~ASSourceIterator() = default;
 	virtual streamoff getPeekStart() const = 0;
 	virtual int getStreamLength() const = 0;
 	virtual bool hasMoreLines() const = 0;
-	virtual string nextLine(bool emptyLineWasDeleted = false) = 0;
+	virtual string nextLine(bool emptyLineWasDeleted) = 0;
 	virtual string peekNextLine() = 0;
 	virtual void peekReset() = 0;
 	virtual streamoff tellg() = 0;
@@ -296,10 +296,10 @@ class ASBase : protected ASResource
 {
 private:
 	// all variables should be set by the "init" function
-	int baseFileType;      // a value from enum FileType
+	int baseFileType = C_TYPE;      // a value from enum FileType
 
 protected:
-	ASBase() : baseFileType(C_TYPE) { }
+	ASBase() = default;
 
 protected:  // inline functions
 	void init(int fileTypeArg) { baseFileType = fileTypeArg; }
@@ -333,7 +333,12 @@ class ASBeautifier : protected ASBase
 public:
 	ASBeautifier();
 	virtual ~ASBeautifier();
+	ASBeautifier(const ASBeautifier& other);
+	ASBeautifier& operator=(ASBeautifier const&) = delete;
+	ASBeautifier(ASBeautifier&&)                 = delete;
+	ASBeautifier& operator=(ASBeautifier&&)      = delete;
 	virtual void init(ASSourceIterator* iter);
+
 	virtual string beautify(const string& originalLine);
 	void setCaseIndent(bool state);
 	void setClassIndent(bool state);
@@ -406,9 +411,6 @@ protected:
 	bool isInIndentablePreproc;
 
 private:  // functions
-	ASBeautifier(const ASBeautifier& other);     // inline functions
-	ASBeautifier& operator=(ASBeautifier&);      // not to be implemented
-
 	void adjustObjCMethodDefinitionIndentation(const string& line_);
 	void adjustObjCMethodCallIndentation(const string& line_);
 	void adjustParsedLineIndentation(size_t iPrelim, bool isInExtraHeaderIndent);
@@ -580,8 +582,7 @@ private:  // variables
 class ASEnhancer : protected ASBase
 {
 public:  // functions
-	ASEnhancer();
-	virtual ~ASEnhancer();
+	ASEnhancer() = default;
 	void init(int, int, int, bool, bool, bool, bool, bool, bool, bool,
 	          vector<const pair<const string, const string>* >*);
 	void enhance(string& line, bool isInNamespace, bool isInPreprocessor, bool isInSQL);
@@ -657,9 +658,14 @@ class ASFormatter : public ASBeautifier
 public:	// functions
 	ASFormatter();
 	virtual ~ASFormatter();
-	virtual void init(ASSourceIterator* si);
-	virtual bool hasMoreLines() const;
-	virtual string nextLine();
+	ASFormatter(const ASFormatter&)            = delete;
+	ASFormatter& operator=(ASFormatter const&) = delete;
+	ASFormatter(ASFormatter&&)                 = delete;
+	ASFormatter& operator=(ASFormatter&&)      = delete;
+	void init(ASSourceIterator* si) override;
+
+	bool hasMoreLines() const;
+	string nextLine();
 	LineEndFormat getLineEndFormat() const;
 	bool getIsLineReady() const;
 	void setFormattingStyle(FormatStyle style);
@@ -722,8 +728,6 @@ public:	// functions
 
 
 private:  // functions
-	ASFormatter(const ASFormatter& copy);       // not to be implemented
-	ASFormatter& operator=(ASFormatter&);       // not to be implemented
 	template<typename T> void deleteContainer(T& container);
 	template<typename T> void initContainer(T& container, T value);
 	char peekNextChar() const;
@@ -752,7 +756,7 @@ private:  // functions
 	bool isMultiStatementLine() const;
 	bool isNextWordSharpNonParenHeader(int startChar) const;
 	bool isNonInStatementArrayBrace() const;
-	bool isNumericVariable(string word) const;
+	bool isNumericVariable(const string& word) const;
 	bool isOkToSplitFormattedLine();
 	bool isPointerOrReference() const;
 	bool isPointerOrReferenceCentered() const;
